Move ministers index spec helpers to support directory
This will allow them to be used in multiple test files.
